moves tests into separate files
